mapcore-chat-backend/flask-chat-backend/requirements.txt
flask>=2.0,<3.0
werkzeug<2.3.0
Flask-Cors==3.0.10
pytest==7.1.2
requests==2.26.0
google-generativeai>=0.3.1
python-dotenv>=0.21.0
# RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
chromadb>=0.4.0
pypdf>=3.0.0
python-docx>=0.8.11